+++ src/cryptoadvance/specter/merkleblock.py
@@ -1,9 +1,8 @@
-# # Code adopted from https://github.com/jimmysong/pb-exercises/
+# Code adopted from https://github.com/jimmysong/pb-exercises/
 import hashlib
 import math
 
 from io import BytesIO
-from unittest import TestCase
 
 
 def hash256(s):
@@ -31,6 +30,41 @@ def merkle_parent(hash1, hash2):
     '''Takes the binary hashes and calculates the hash256'''
     # return the hash256 of hash1 + hash2
     return hash256(hash1 + hash2)
+
+
+def merkle_parent_level(hashes):
+    '''Takes a list of binary hashes and returns a list that's half
+    the length'''
+    # if the list has exactly 1 element raise an error
+    if len(hashes) == 1:
+        raise RuntimeError('Cannot take a parent level with only 1 item')
+    # if the list has an odd number of elements, duplicate the last one
+    #       and put it at the end so it has an even number of elements
+    if len(hashes) % 2 == 1:
+        hashes.append(hashes[-1])
+    # initialize parent level
+    parent_level = []
+    # loop over every pair (use: for i in range(0, len(hashes), 2))
+    for i in range(0, len(hashes), 2):
+        # get the merkle parent of i and i+1 hashes
+        parent = merkle_parent(hashes[i], hashes[i + 1])
+        # append parent to parent level
+        parent_level.append(parent)
+    # return parent level
+    return parent_level
+
+
+def merkle_root(hashes):
+    '''Takes a list of binary hashes and returns the merkle root
+    '''
+    # current level starts as hashes
+    current_level = hashes
+    # loop until there's exactly 1 element
+    while len(current_level) > 1:
+        # current level becomes the merkle parent level
+        current_level = merkle_parent_level(current_level)
+    # return the 1st item of current_level
+    return current_level[0]
 
 
 def little_endian_to_int(b):
@@ -190,100 +224,6 @@ class Block:
         return root[::-1] == self.merkle_root
 
 
-class BlockTest(TestCase):
-
-    def test_parse(self):
-        block_raw = bytes.fromhex('020000208ec39428b17323fa0ddec8e887b4a7c53b8c0a0a220cfd0000000000000000005b0750fce0a889502d40508d39576821155e9c9e3f5c3157f961db38fd8b25be1e77a759e93c0118a4ffd71d')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertEqual(block.version, 0x20000002)
-        want = bytes.fromhex('000000000000000000fd0c220a0a8c3bc5a7b487e8c8de0dfa2373b12894c38e')
-        self.assertEqual(block.prev_block, want)
-        want = bytes.fromhex('be258bfd38db61f957315c3f9e9c5e15216857398d50402d5089a8e0fc50075b')
-        self.assertEqual(block.merkle_root, want)
-        self.assertEqual(block.timestamp, 0x59a7771e)
-        self.assertEqual(block.bits, bytes.fromhex('e93c0118'))
-        self.assertEqual(block.nonce, bytes.fromhex('a4ffd71d'))
-
-    def test_serialize(self):
-        block_raw = bytes.fromhex('020000208ec39428b17323fa0ddec8e887b4a7c53b8c0a0a220cfd0000000000000000005b0750fce0a889502d40508d39576821155e9c9e3f5c3157f961db38fd8b25be1e77a759e93c0118a4ffd71d')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertEqual(block.serialize(), block_raw)
-
-    def test_hash(self):
-        block_raw = bytes.fromhex('020000208ec39428b17323fa0ddec8e887b4a7c53b8c0a0a220cfd0000000000000000005b0750fce0a889502d40508d39576821155e9c9e3f5c3157f961db38fd8b25be1e77a759e93c0118a4ffd71d')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertEqual(block.hash(), bytes.fromhex('0000000000000000007e9e4c586439b0cdbe13b1370bdd9435d76a644d047523'))
-
-    def test_bip9(self):
-        block_raw = bytes.fromhex('020000208ec39428b17323fa0ddec8e887b4a7c53b8c0a0a220cfd0000000000000000005b0750fce0a889502d40508d39576821155e9c9e3f5c3157f961db38fd8b25be1e77a759e93c0118a4ffd71d')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertTrue(block.bip9())
-        block_raw = bytes.fromhex('0400000039fa821848781f027a2e6dfabbf6bda920d9ae61b63400030000000000000000ecae536a304042e3154be0e3e9a8220e5568c3433a9ab49ac4cbb74f8df8e8b0cc2acf569fb9061806652c27')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertFalse(block.bip9())
-
-    def test_bip91(self):
-        block_raw = bytes.fromhex('1200002028856ec5bca29cf76980d368b0a163a0bb81fc192951270100000000000000003288f32a2831833c31a25401c52093eb545d28157e200a64b21b3ae8f21c507401877b5935470118144dbfd1')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertTrue(block.bip91())
-        block_raw = bytes.fromhex('020000208ec39428b17323fa0ddec8e887b4a7c53b8c0a0a220cfd0000000000000000005b0750fce0a889502d40508d39576821155e9c9e3f5c3157f961db38fd8b25be1e77a759e93c0118a4ffd71d')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertFalse(block.bip91())
-
-    def test_bip141(self):
-        block_raw = bytes.fromhex('020000208ec39428b17323fa0ddec8e887b4a7c53b8c0a0a220cfd0000000000000000005b0750fce0a889502d40508d39576821155e9c9e3f5c3157f961db38fd8b25be1e77a759e93c0118a4ffd71d')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertTrue(block.bip141())
-        block_raw = bytes.fromhex('0000002066f09203c1cf5ef1531f24ed21b1915ae9abeb691f0d2e0100000000000000003de0976428ce56125351bae62c5b8b8c79d8297c702ea05d60feabb4ed188b59c36fa759e93c0118b74b2618')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertFalse(block.bip141())
-
-    def test_target(self):
-        block_raw = bytes.fromhex('020000208ec39428b17323fa0ddec8e887b4a7c53b8c0a0a220cfd0000000000000000005b0750fce0a889502d40508d39576821155e9c9e3f5c3157f961db38fd8b25be1e77a759e93c0118a4ffd71d')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertEqual(block.target(), 0x13ce9000000000000000000000000000000000000000000)
-        self.assertEqual(int(block.difficulty()), 888171856257)
-
-    def test_check_pow(self):
-        block_raw = bytes.fromhex('04000000fbedbbf0cfdaf278c094f187f2eb987c86a199da22bbb20400000000000000007b7697b29129648fa08b4bcd13c9d5e60abb973a1efac9c8d573c71c807c56c3d6213557faa80518c3737ec1')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertTrue(block.check_pow())
-        block_raw = bytes.fromhex('04000000fbedbbf0cfdaf278c094f187f2eb987c86a199da22bbb20400000000000000007b7697b29129648fa08b4bcd13c9d5e60abb973a1efac9c8d573c71c807c56c3d6213557faa80518c3737ec0')
-        stream = BytesIO(block_raw)
-        block = Block.parse_header(stream)
-        self.assertFalse(block.check_pow())
-
-    def test_validate_merkle_root(self):
-        hashes_hex = [
-            'f54cb69e5dc1bd38ee6901e4ec2007a5030e14bdd60afb4d2f3428c88eea17c1',
-            'c57c2d678da0a7ee8cfa058f1cf49bfcb00ae21eda966640e312b464414731c1',
-            'b027077c94668a84a5d0e72ac0020bae3838cb7f9ee3fa4e81d1eecf6eda91f3',
-            '8131a1b8ec3a815b4800b43dff6c6963c75193c4190ec946b93245a9928a233d',
-            'ae7d63ffcb3ae2bc0681eca0df10dda3ca36dedb9dbf49e33c5fbe33262f0910',
-            '61a14b1bbdcdda8a22e61036839e8b110913832efd4b086948a6a64fd5b3377d',
-            'fc7051c8b536ac87344c5497595d5d2ffdaba471c73fae15fe9228547ea71881',
-            '77386a46e26f69b3cd435aa4faac932027f58d0b7252e62fb6c9c2489887f6df',
-            '59cbc055ccd26a2c4c4df2770382c7fea135c56d9e75d3f758ac465f74c025b8',
-            '7c2bf5687f19785a61be9f46e031ba041c7f93e2b7e9212799d84ba052395195',
-            '08598eebd94c18b0d59ac921e9ba99e2b8ab7d9fccde7d44f2bd4d5e2e726d2e',
-            'f0bb99ef46b029dd6f714e4b12a7d796258c48fee57324ebdc0bbc4700753ab1',
-        ]
-        hashes = [bytes.fromhex(x) for x in hashes_hex]
-        stream = BytesIO(bytes.fromhex('00000020fcb19f7895db08cadc9573e7915e3919fb76d59868a51d995201000000000000acbcab8bcc1af95d8d563b77d24c3d19b18f1486383d75a5085c4e86c86beed691cfa85916ca061a00000000'))
-        block = Block.parse_header(stream)
-        block.tx_hashes = hashes
-        self.assertTrue(block.validate_merkle_root())
 
 
 class MerkleTree:
@@ -418,56 +358,6 @@ class MerkleTree:
                 raise RuntimeError('flag bits not all consumed')
 
 
-class MerkleTreeTest(TestCase):
-
-    def test_init(self):
-        tree = MerkleTree(9)
-        self.assertEqual(len(tree.nodes[0]), 1)
-        self.assertEqual(len(tree.nodes[1]), 2)
-        self.assertEqual(len(tree.nodes[2]), 3)
-        self.assertEqual(len(tree.nodes[3]), 5)
-        self.assertEqual(len(tree.nodes[4]), 9)
-
-    def test_populate_tree_1(self):
-        hex_hashes = [
-            "9745f7173ef14ee4155722d1cbf13304339fd00d900b759c6f9d58579b5765fb",
-            "5573c8ede34936c29cdfdfe743f7f5fdfbd4f54ba0705259e62f39917065cb9b",
-            "82a02ecbb6623b4274dfcab82b336dc017a27136e08521091e443e62582e8f05",
-            "507ccae5ed9b340363a0e6d765af148be9cb1c8766ccc922f83e4ae681658308",
-            "a7a4aec28e7162e1e9ef33dfa30f0bc0526e6cf4b11a576f6c5de58593898330",
-            "bb6267664bd833fd9fc82582853ab144fece26b7a8a5bf328f8a059445b59add",
-            "ea6d7ac1ee77fbacee58fc717b990c4fcccf1b19af43103c090f601677fd8836",
-            "457743861de496c429912558a106b810b0507975a49773228aa788df40730d41",
-            "7688029288efc9e9a0011c960a6ed9e5466581abf3e3a6c26ee317461add619a",
-            "b1ae7f15836cb2286cdd4e2c37bf9bb7da0a2846d06867a429f654b2e7f383c9",
-            "9b74f89fa3f93e71ff2c241f32945d877281a6a50a6bf94adac002980aafe5ab",
-            "b3a92b5b255019bdaf754875633c2de9fec2ab03e6b8ce669d07cb5b18804638",
-            "b5c0b915312b9bdaedd2b86aa2d0f8feffc73a2d37668fd9010179261e25e263",
-            "c9d52c5cb1e557b92c84c52e7c4bfbce859408bedffc8a5560fd6e35e10b8800",
-            "c555bc5fc3bc096df0a0c9532f07640bfb76bfe4fc1ace214b8b228a1297a4c2",
-            "f9dbfafc3af3400954975da24eb325e326960a25b87fffe23eef3e7ed2fb610e",
-        ]
-        tree = MerkleTree(len(hex_hashes))
-        hashes = [bytes.fromhex(h) for h in hex_hashes]
-        tree.populate_tree([1] * 31, hashes)
-        root = '597c4bafe3832b17cbbabe56f878f4fc2ad0f6a402cee7fa851a9cb205f87ed1'
-        self.assertEqual(tree.root().hex(), root)
-
-    def test_populate_tree_2(self):
-        hex_hashes = [
-            '42f6f52f17620653dcc909e58bb352e0bd4bd1381e2955d19c00959a22122b2e',
-            '94c3af34b9667bf787e1c6a0a009201589755d01d02fe2877cc69b929d2418d4',
-            '959428d7c48113cb9149d0566bde3d46e98cf028053c522b8fa8f735241aa953',
-            'a9f27b99d5d108dede755710d4a1ffa2c74af70b4ca71726fa57d68454e609a2',
-            '62af110031e29de1efcad103b3ad4bec7bdcf6cb9c9f4afdd586981795516577',
-        ]
-        tree = MerkleTree(len(hex_hashes))
-        hashes = [bytes.fromhex(h) for h in hex_hashes]
-        tree.populate_tree([1] * 11, hashes)
-        root = 'a8e8bd023169b81bc56854137a135b97ef47a6a7237f4c6e037baed16285a5ab'
-        self.assertEqual(tree.root().hex(), root)
-
-
 class MerkleBlock:
     command = b'merkleblock'
 
@@ -532,52 +422,6 @@ class MerkleBlock:
             return []
         else:
             return self.merkle_tree.proved_txs
-
-
-class MerkleBlockTest(TestCase):
-
-    def test_parse(self):
-        hex_merkle_block = '00000020df3b053dc46f162a9b00c7f0d5124e2676d47bbe7c5d0793a500000000000000ef445fef2ed495c275892206ca533e7411907971013ab83e3b47bd0d692d14d4dc7c835b67d8001ac157e670bf0d00000aba412a0d1480e370173072c9562becffe87aa661c1e4a6dbc305d38ec5dc088a7cf92e6458aca7b32edae818f9c2c98c37e06bf72ae0ce80649a38655ee1e27d34d9421d940b16732f24b94023e9d572a7f9ab8023434a4feb532d2adfc8c2c2158785d1bd04eb99df2e86c54bc13e139862897217400def5d72c280222c4cbaee7261831e1550dbb8fa82853e9fe506fc5fda3f7b919d8fe74b6282f92763cef8e625f977af7c8619c32a369b832bc2d051ecd9c73c51e76370ceabd4f25097c256597fa898d404ed53425de608ac6bfe426f6e2bb457f1c554866eb69dcb8d6bf6f880e9a59b3cd053e6c7060eeacaacf4dac6697dac20e4bd3f38a2ea2543d1ab7953e3430790a9f81e1c67f5b58c825acf46bd02848384eebe9af917274cdfbb1a28a5d58a23a17977def0de10d644258d9c54f886d47d293a411cb6226103b55635'
-        mb = MerkleBlock.parse(BytesIO(bytes.fromhex(hex_merkle_block)))
-        version = 0x20000000
-        self.assertEqual(mb.header.version, version)
-        merkle_root_hex = 'ef445fef2ed495c275892206ca533e7411907971013ab83e3b47bd0d692d14d4'
-        merkle_root = bytes.fromhex(merkle_root_hex)[::-1]
-        self.assertEqual(mb.header.merkle_root, merkle_root)
-        prev_block_hex = 'df3b053dc46f162a9b00c7f0d5124e2676d47bbe7c5d0793a500000000000000'
-        prev_block = bytes.fromhex(prev_block_hex)[::-1]
-        self.assertEqual(mb.header.prev_block, prev_block)
-        timestamp = little_endian_to_int(bytes.fromhex('dc7c835b'))
-        self.assertEqual(mb.header.timestamp, timestamp)
-        bits = bytes.fromhex('67d8001a')
-        self.assertEqual(mb.header.bits, bits)
-        nonce = bytes.fromhex('c157e670')
-        self.assertEqual(mb.header.nonce, nonce)
-        total = little_endian_to_int(bytes.fromhex('bf0d0000'))
-        self.assertEqual(mb.total, total)
-        hex_hashes = [
-            'ba412a0d1480e370173072c9562becffe87aa661c1e4a6dbc305d38ec5dc088a',
-            '7cf92e6458aca7b32edae818f9c2c98c37e06bf72ae0ce80649a38655ee1e27d',
-            '34d9421d940b16732f24b94023e9d572a7f9ab8023434a4feb532d2adfc8c2c2',
-            '158785d1bd04eb99df2e86c54bc13e139862897217400def5d72c280222c4cba',
-            'ee7261831e1550dbb8fa82853e9fe506fc5fda3f7b919d8fe74b6282f92763ce',
-            'f8e625f977af7c8619c32a369b832bc2d051ecd9c73c51e76370ceabd4f25097',
-            'c256597fa898d404ed53425de608ac6bfe426f6e2bb457f1c554866eb69dcb8d',
-            '6bf6f880e9a59b3cd053e6c7060eeacaacf4dac6697dac20e4bd3f38a2ea2543',
-            'd1ab7953e3430790a9f81e1c67f5b58c825acf46bd02848384eebe9af917274c',
-            'dfbb1a28a5d58a23a17977def0de10d644258d9c54f886d47d293a411cb62261',
-        ]
-        hashes = [bytes.fromhex(h)[::-1] for h in hex_hashes]
-        self.assertEqual(mb.hashes, hashes)
-        flags = bytes.fromhex('b55635')
-        self.assertEqual(mb.flags, flags)
-
-    def test_is_valid(self):
-        hex_merkle_block = '00000020df3b053dc46f162a9b00c7f0d5124e2676d47bbe7c5d0793a500000000000000ef445fef2ed495c275892206ca533e7411907971013ab83e3b47bd0d692d14d4dc7c835b67d8001ac157e670bf0d00000aba412a0d1480e370173072c9562becffe87aa661c1e4a6dbc305d38ec5dc088a7cf92e6458aca7b32edae818f9c2c98c37e06bf72ae0ce80649a38655ee1e27d34d9421d940b16732f24b94023e9d572a7f9ab8023434a4feb532d2adfc8c2c2158785d1bd04eb99df2e86c54bc13e139862897217400def5d72c280222c4cbaee7261831e1550dbb8fa82853e9fe506fc5fda3f7b919d8fe74b6282f92763cef8e625f977af7c8619c32a369b832bc2d051ecd9c73c51e76370ceabd4f25097c256597fa898d404ed53425de608ac6bfe426f6e2bb457f1c554866eb69dcb8d6bf6f880e9a59b3cd053e6c7060eeacaacf4dac6697dac20e4bd3f38a2ea2543d1ab7953e3430790a9f81e1c67f5b58c825acf46bd02848384eebe9af917274cdfbb1a28a5d58a23a17977def0de10d644258d9c54f886d47d293a411cb6226103b55635'
-        mb = MerkleBlock.parse(BytesIO(bytes.fromhex(hex_merkle_block)))
-        self.assertTrue(mb.is_valid())
-        want = '6122b61c413a297dd486f8549c8d2544d610def0de7779a1238ad5a5281abbdf'
-        self.assertEqual(mb.proved_txs()[0].hex(), want)
 
 
 def is_valid_merkle_proof(proof_hex, target_tx_hex, target_block_hash_hex, target_merkle_root_hex=None):
